+++ plugins/ilert/src/components/ILertCard/ILertCardEmptyState.tsx
@@ -41,7 +41,8 @@ const useStyles = makeStyles<BackstageTheme>(theme => ({
   code: {
     borderRadius: 6,
     margin: theme.spacing(2, 0),
-    background: theme.palette.type === 'dark' ? '#444' : '#fff',
+    background:
+      theme.palette.type === 'dark' ? '#444' : theme.palette.common.white,
   },
   header: {
     display: 'inline-block',