+++ plugins/scaffolder-react/src/next/components/TaskSteps/TaskBorder.tsx
@@ -28,8 +28,6 @@ const useStyles = makeStyles((theme: BackstageTheme) => ({
 
 /**
  * The visual progress of the task event stream
- *
- * @public
  */
 export const TaskBorder = (props: {
   isComplete: boolean;