+++ plugins/sonarqube/src/api/SonarQubeClient.ts
@@ -28,15 +28,12 @@ export class SonarQubeClient implements SonarQubeApi {
   discoveryApi: DiscoveryApi;
   identityApi: IdentityApi;
 
-  constructor({
-    discoveryApi,
-    identityApi,
-  }: {
+  constructor(options: {
     discoveryApi: DiscoveryApi;
     identityApi: IdentityApi;
   }) {
-    this.discoveryApi = discoveryApi;
-    this.identityApi = identityApi;
+    this.discoveryApi = options.discoveryApi;
+    this.identityApi = options.identityApi;
   }
 
   private async callApi<T>(