+++ plugins/catalog-backend-module-gerrit/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @backstage/plugin-catalog-backend-module-gerrit
 
+## 0.1.11-next.1
+
+### Patch Changes
+
+- 52b0022dab7: Updated dependency `msw` to `^1.0.0`.
+- Updated dependencies
+  - @backstage/errors@1.1.5-next.0
+  - @backstage/backend-common@0.18.3-next.1
+  - @backstage/integration@1.4.3-next.0
+  - @backstage/plugin-catalog-backend@1.8.0-next.1
+  - @backstage/backend-plugin-api@0.4.1-next.1
+  - @backstage/backend-tasks@0.4.4-next.1
+  - @backstage/config@1.0.7-next.0
+  - @backstage/catalog-model@1.2.1-next.1
+  - @backstage/plugin-catalog-node@1.3.4-next.1
+
 ## 0.1.11-next.0
 
 ### Patch Changes