+++ plugins/search-react/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @backstage/plugin-search-react
 
+## 1.5.1-next.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/core-components@0.12.5-next.1
+  - @backstage/core-plugin-api@1.4.1-next.1
+  - @backstage/theme@0.2.18-next.0
+  - @backstage/types@1.0.2
+  - @backstage/version-bridge@1.0.3
+  - @backstage/plugin-search-common@1.2.2-next.0
+
 ## 1.5.1-next.0
 
 ### Patch Changes